chore: delete questionparser
It was not really that valuable, more for me to learn how to write a parser. The problem with it is that it wasn’t recursive, so couldn’t render *hello _there_* as an italic there inside a bold sentence.

I tried with the “simple-text-parser” npm module but that didn’t work either. I decided it’s not worth it to try and support this functionality.